% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pollencast.R
\name{pollencast}
\alias{pollencast}
\title{Retrieve pollen count forecast}
\usage{
pollencast(zip = NULL)
}
\arguments{
\item{zip}{five digit zip code}
}
\description{
Retrieve pollen count forecast
}
\examples{
df  <- pollencast(zip = 20238)
df2 <- pollencast(zip = "20238")
}

