% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollencast.R
\name{pollencast}
\alias{pollencast}
\title{Retrieve pollen count forecast}
\usage{
pollencast(zip = NULL, api_key = NULL, forecast_days = 1,
  metric = "false", language_code = "en-us", user_agent = NULL)
}
\arguments{
\item{zip}{five digit zip code}

\item{api_key}{API key string from Accuweather}

\item{forecast_days}{desired number of forecast days; 1 by default, accepts 1, 5, 10, and 15}

\item{metric}{Use metric system? defaults to "false"}

\item{user_agent}{user agent string (optional); otherwise uses httr's default user agent}

\item{language}{defaults to US English, see Accuweather API docs for supported languages}
}
\description{
Retrieve pollen count forecast
}
\examples{
df  <- pollencast(zip = 20238)
df2 <- pollencast(zip = "20238")
}
